% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_caption.r
\name{bn_cap}
\alias{bn_cap}
\alias{bn_c}
\title{table caption}
\usage{
bn_cap(caption, label = NULL)

bn_c(caption, label = NULL)
}
\arguments{
\item{caption}{character; the caption}

\item{label}{character; the label. By default, this will be the chunk label,
most of the time you do not want to change this}
}
\description{
get bookdown enumerable table caption with label from the chunk label as
    default. The caption provided will be inserted after
    "(\\#tab:<chunk label>)"
}
\section{Functions}{
\itemize{
\item \code{bn_c}: alias for \code{bn_cap}
}}

