% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_table_fix.r
\name{bn_double_table_fix}
\alias{bn_double_table_fix}
\alias{bn_double_table_fix_docx}
\alias{bn_double_table_fix_html}
\title{fix compiled file}
\usage{
bn_double_table_fix(input, keep_input = TRUE, prefix = "raw-")

bn_double_table_fix_docx(input, keep_input = TRUE, prefix = "raw-")

bn_double_table_fix_html(input, keep_input = TRUE, prefix = "raw-")
}
\arguments{
\item{input}{character; name of docx file in the working directory}

\item{keep_input}{logical; should input be kept?}

\item{prefix}{character; if input is kept, the prefix to use}
}
\value{
NULL; one or two files are written/manipulated as a side effect
}
\description{
Some table code generators (I'm thinking specifically of huxtable) have to
    have caption set to "Table: (\#<label>)" in order for automatic
    enumeration with bookdown to work for "all" formats (html, pdf, docx),
    with the side effect that tables will have captions starting with
    "Table: Table". This function simply substitutes "Table: Table" with
    "Table:". Note: this function is for docx and html, for pdf you need a
    'hook' - run \code{bn_double_table_fix(".pdf")} for more info
}
\section{Functions}{
\itemize{
\item \code{bn_double_table_fix_docx}: fix double table instance in docx file

\item \code{bn_double_table_fix_html}: fix double table instance in html file
}}

