% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_table_fix.r
\name{bn_double_table_latex_hook}
\alias{bn_double_table_latex_hook}
\title{fix pre-compiled file}
\usage{
bn_double_table_latex_hook(x = NULL)
}
\arguments{
\item{x}{argument used by internal knitr process; set to \code{NULL}
(default) to get a helpful message}
}
\description{
bookdown with pdf output are in some instances subject to the same problems
    described in \code{\link{bn_double_table_fix}}. This 'hook' tries to
    remove that problem. Insert this into a relevant setup chunk:
    \code{knit_hooks$set(document = bn_double_table_latex_hook)}
}
