% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.r
\name{bn_if_X}
\alias{bn_if_X}
\alias{bn_if_docx}
\alias{bn_if_latex}
\alias{bn_if_html}
\title{conditional code}
\usage{
bn_if_X(X, cc)

bn_if_docx(cc)

bn_if_latex(cc)

bn_if_html(cc)
}
\arguments{
\item{X}{format}

\item{cc}{character; code conditional on format being X}
}
\description{
if format is X, insert conditional code specified by 'cc'
}
\section{Functions}{
\itemize{
\item \code{bn_if_docx}: code conditional on format being docx

\item \code{bn_if_latex}: code conditional on format being latex

\item \code{bn_if_html}: code conditional on format being html
}}

