% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.r
\name{bn_is_X}
\alias{bn_is_X}
\alias{bn_is_docx}
\alias{bn_is_latex}
\alias{bn_is_html}
\title{test for output format}
\usage{
bn_is_X(X = NULL)

bn_is_docx()

bn_is_latex()

bn_is_html()
}
\arguments{
\item{X}{format to test for}
}
\description{
check if knitr option 'rmarkdown.pandoc.to' is a given value
}
\section{Functions}{
\itemize{
\item \code{bn_is_docx}: test for docx

\item \code{bn_is_latex}: test for latex

\item \code{bn_is_html}: test for html
}}

