% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.r
\name{bn_newpage}
\alias{bn_newpage}
\alias{bn_newpage_docx}
\alias{bn_newpage_latex}
\title{set new page}
\usage{
bn_newpage(format = c("docx", "latex"))

bn_newpage_docx()

bn_newpage_latex()
}
\arguments{
\item{format}{character vector of formats}
}
\description{
insert relevant code for 'newpage' in all formats specified by argument
    'format' (support for latex and docx)
}
\section{Functions}{
\itemize{
\item \code{bn_newpage_docx}: newpage code for docx

\item \code{bn_newpage_latex}: newpage code for latex
}}

